// src/app/partials/board/board.enum.ts
export enum TypeSquare {
  Start = 'Start',
  Normal = 'Normal',
  AvantZen = 'AvantZen',
  BatallZen = 'BatallZen',
  RetroZen = 'RetroZen',
  SuertZen = 'SuertZen',
  Finish = 'Finish',
}

export enum CSS_PROPERTIES {
  playerW = 20,
  playerH = 20,
  space = 5,
  border = 5,
}

enum TipoBatalla {
  Dados = 'Batalla por dados',
  Bully = 'Bully',
  Trivial = 'Trivial',
}

enum TipoSuertZen {
  Ruleta = 'Ruleta',
  DadoExtra = 'Dado extra',
  TurnoExtra = 'Turno extra',
  Trivial = 'Trivial',
}

enum TipoRetroZen {
  Ruleta = 'Ruleta',
  DadoExtra = 'Dado extra',
  TurnoExtra = 'Turno extra',
  Trivial = 'Trivial',
}

enum TipoAvantZen {
  Ruleta = 'Ruleta',
  DadoExtra = 'Dado extra',
  TurnoExtra = 'Turno extra',
  Trivial = 'Trivial',
}